requirements.txt
streamlit
pypdf2
langchain
langchain-community
pinecone-client
sentence-transformers
groq
python-dotenv
